fix: correctness, fault-tolerance and dead-code pass (skills review)
Audio correctness:
- q15_mul: widen intermediate multiply to int64 to prevent signed overflow
  when pre-reverb accumulator (up to ±50×32767) meets kDry coefficient
- mixer end-ramp: use (samples_left << 15) / VOICE_END_RAMP_SAMPLES so the
  formula is correct for any ramp length, not just powers of 2
- reverb: replace magic 32767 with Q15_UNITY in wet/dry split

Logic bugs:
- auto_drummer: advance_until now breaks instead of incrementing idx past the
  max_trig budget, preventing silent event drops at loop boundaries
- looper: start_playing() with record_pos_==0 now calls stop() instead of
  entering Playing state with loop_length_==0 (stuck-silent bug)

Fault tolerance:
- main: I2S init failure halts with vTaskDelay loop instead of continuing
  into app_task with an uninitialised driver

Debt / dead code:
- defines.hpp: add BPM_COUNTRY {85,100,120}; use all BPM_* macros in kBpmTable
- hal_native: remove unused set_capacity() / capacity_ (dead since write_nonblocking removed)

+++ src/util/dsp.hpp
@@ -18,7 +18,7 @@ inline int16_t sat_i16(int32_t v)
 
 inline int32_t q15_mul(int32_t a, int32_t b)
 {
-    return (a * b) >> 15;
+    return static_cast<int32_t>((static_cast<int64_t>(a) * b) >> 15);
 }
 
 } // namespace util
